feat: validate publisher deletion with books

+++ src/main/java/com/library/api/services/impl/PublisherServiceImpl.java
@@ -5,6 +5,7 @@ import com.library.api.dtos.publisher.PublisherRequestDTO;
 import com.library.api.dtos.publisher.PublisherResponseDTO;
 import com.library.api.entities.Publisher;
 import com.library.api.mappers.PublisherMapper;
+import com.library.api.repositories.BookRepository;
 import com.library.api.repositories.PublisherRepository;
 import com.library.api.services.PublisherService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -19,6 +20,8 @@ import org.springframework.web.server.ResponseStatusException;
 public class PublisherServiceImpl implements PublisherService {
     @Autowired
     private PublisherRepository publisherRepository;
+    @Autowired
+    private BookRepository bookRepository;
     @Autowired
     private PublisherMapper publisherMapper;
 
@@ -73,9 +76,10 @@ public class PublisherServiceImpl implements PublisherService {
 
     @Override
     public void deletePublisher(Long id) {
-        if (!publisherRepository.existsById(id)) {
+        if (!publisherRepository.existsById(id))
             throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Editora não encontrada");
-        }
+        if (bookRepository.existsByPublisherId(id))
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Existem livros cadastrados com essa editora");
 
         publisherRepository.deleteById(id);
     }
